refactor: replace with UserController.deleteFavorite
